Split MainNavigation.kt into MainNavHost & NavigationRouter
