+++ app/src/main/java/srimani7/apps/feedfly/MainActivity.kt
@@ -15,6 +15,7 @@ import srimani7.apps.feedfly.core.design.TheSecretDairyTheme
 import srimani7.apps.feedfly.core.preferences.UserSettingsRepo
 import srimani7.apps.feedfly.core.preferences.model.AppTheme
 import srimani7.apps.feedfly.core.preferences.model.ThemePreference
+import srimani7.apps.feedfly.navigation.MainNavHost
 import srimani7.apps.feedfly.navigation.URL_REGEX
 
 class MainActivity : ComponentActivity() {
@@ -42,7 +43,7 @@ class MainActivity : ComponentActivity() {
                 )
             }
             TheSecretDairyTheme(darkTheme, dynamicColor = themePreference.useDynamicTheme) {
-                MainNavigation(feedUrl)
+                MainNavHost(feedUrl)
             }
         }
