+++ app/src/main/java/srimani7/apps/feedfly/NavigationRouter.kt
@@ -1,0 +1,11 @@
+package srimani7.apps.feedfly
+
+import srimani7.apps.feedfly.navigation.Screen
+
+object NavigationRouter {
+    fun newFeedRoute() = Screen.InsertFeedScreen.destination
+    fun groupOverviewScreen(name: String) = Screen.GroupOverviewScreen.destination + "/${name}"
+    fun articlesScreenRoute(id: Long) = Screen.ArticlesScreen.destination + "/${id}"
+    fun privateSpaceRoute() = Screen.PrivateSpaceScreen.destination
+    fun labelRoute(id: Long) = "labels/$id"
+}
