+++ app/src/main/java/srimani7/apps/feedfly/navigation/MainNavHost.kt
@@ -1,4 +1,4 @@
-package srimani7.apps.feedfly
+package srimani7.apps.feedfly.navigation
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -6,10 +6,6 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -20,29 +16,20 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.navigation.NavController
 import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.dialog
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
+import srimani7.apps.feedfly.NavigationRouter
 import srimani7.apps.feedfly.feature.labels.ui.LabelOverviewScaffold
 import srimani7.apps.feedfly.feature.labels.ui.LabelsScaffold
-import srimani7.apps.feedfly.navigation.ArticlesScreen
-import srimani7.apps.feedfly.navigation.ChangeArticleLabelDialog
-import srimani7.apps.feedfly.navigation.GroupOverviewScreen
-import srimani7.apps.feedfly.navigation.HomeScreen
-import srimani7.apps.feedfly.navigation.NewFeedScreen
-import srimani7.apps.feedfly.navigation.PrivateSpaceScreen
-import srimani7.apps.feedfly.navigation.RemoveArticlesScreen
-import srimani7.apps.feedfly.navigation.Screen
-import srimani7.apps.feedfly.navigation.SettingsScreen
 import srimani7.apps.feedfly.viewmodel.HomeViewModal
 import srimani7.apps.feedfly.viewmodel.SettingsViewModel
 
 @Composable
-fun MainNavigation(addLink: String?) {
+fun MainNavHost(addLink: String?) {
     val homeViewModal = viewModel<HomeViewModal>()
     val navController = rememberNavController()
     val deletingState by homeViewModal.deletingStateFlow.collectAsStateWithLifecycle()
@@ -67,7 +54,7 @@ fun MainNavigation(addLink: String?) {
                 val labels by homeViewModal.labels.collectAsStateWithLifecycle(initialValue = emptyList())
                 LabelsScaffold(
                     labelData = labels,
-                    onClick = { id, _ -> navController.navigate(MainNavigation.labelRoute(id)) },
+                    onClick = { id, _ -> navController.navigate(NavigationRouter.labelRoute(id)) },
                     onBackClick = navController::popBackStack,
                     onAddNewLabel = {
                         homeViewModal.addLabel(it)
@@ -149,20 +136,4 @@ fun MainNavigation(addLink: String?) {
     LaunchedEffect(addLink) {
         addLink?.let { navController.navigate(Screen.InsertFeedScreen.destination) }
     }
-}
-
-
-@Composable
-fun BackButton(navController: NavController) {
-    IconButton(onClick = { navController.popBackStack() }) {
-        Icon(Icons.Default.ArrowBack, "back")
-    }
-}
-
-object MainNavigation {
-    fun newFeedRoute() = Screen.InsertFeedScreen.destination
-    fun groupOverviewScreen(name: String) = Screen.GroupOverviewScreen.destination + "/${name}"
-    fun articlesScreenRoute(id: Long) = Screen.ArticlesScreen.destination + "/${id}"
-    fun privateSpaceRoute() = Screen.PrivateSpaceScreen.destination
-    fun labelRoute(id: Long) = "labels/$id"
 }
